Enable enhanced KCFG view in kmir [view|show] (#558)
This PR enables a richer view of the KCFG in both `kcfg show ...` and
`kcfg view ...`. In particular:

- The options class `DisplayOpts` is factored out for both `ShowOpts`
and `ViewOpts`, and the parser is factored out as well.
- Option `kmir prove ... --save-smir ...` now will not delete the SMIR
file when it's done, and will inform you (via `--verbose` or `--debug`
logs) where it saved it.
- Option `kmir show ... --no-full-printer` enables not printing out the
entire configuration every step.
- Option `kmir [show|view] ... --smir-info <path/to/smir.json> ...`
enables augmented output with span maps and function names.
- Adds `symbol` attribute to `ty` and `span` in K so that it's easy to
construct/match them.

To make use of the source maps, you need to keep the `smir.json` around
of a given file. For example:

```
# NOTICE: use of `--verbose --save-mir`
% kmir prove-rs --verbose --save-smir --proof-dir /home/dev/src/mir-semantics/proofs/ kmir/src/tests/integration/data/prove-rs/cmp.rs
...
A BUNCH OF LOG OUTPUT, THEN THE IMPORTANT LINE:
INFO 2025-05-02 02:39:18,046 kmir.cargo - SMIR JSON available at: /home/dev/src/mir-semantics/deps/stable-mir-json/cmp.smir.json
A BUNCH MORE LOG OUTPUT

# Launch the viewer with the SMIR JSON file
# NOTICE: passing in `--smir-info <SMIR JSON>`
% kmir show --proof-dir /home/dev/src/mir-semantics/proofs/ cmp --no-full-printer --smir-info /home/dev/src/mir-semantics-pr/deps/stable-mir-json/cmp.smir.json
```

---------

+++ kmir/src/kmir/cargo.py
@@ -94,11 +94,14 @@ class CargoProject:
         return bin_targets[0]
 
 
-def cargo_get_smir_json(rs_file: Path) -> dict[str, Any]:
+def cargo_get_smir_json(rs_file: Path, save_smir: bool = False) -> dict[str, Any]:
     smir_json_result = SMIR_JSON_DIR / rs_file.with_suffix('.smir.json').name
     run_process_2(['cargo', 'run', '--', '-Zno-codegen', str(rs_file)], cwd=SMIR_JSON_DIR)
     json_smir = json.loads(smir_json_result.read_text())
     _LOGGER.info(f'Loaded: {smir_json_result}')
-    smir_json_result.unlink()
-    _LOGGER.info(f'Deleted: {smir_json_result}')
+    if save_smir:
+        _LOGGER.info(f'SMIR JSON available at: {smir_json_result}')
+    else:
+        smir_json_result.unlink()
+        _LOGGER.info(f'Deleted: {smir_json_result}')
     return json_smir
